# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Logos API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/session-configs/{config_id}": {
      "get": {
        "tags": ["session-configs"],
        "summary": "Get Session Config",
        "operationId": "getSessionConfig",
        "parameters": [
          {
            "name": "config_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Config Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionConfigRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/sessions/{session_id}": {
      "get": {
        "tags": ["sessions"],
        "summary": "Get Session",
        "operationId": "getSession",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentParticipantRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "const": "agent",
            "title": "Type",
            "default": "agent"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "system_prompt": {
            "type": "string",
            "title": "System Prompt"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "updated_at",
          "model",
          "system_prompt"
        ],
        "title": "AgentParticipantRead"
      },
...
      "ContextConfigRead": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/ContextMode"
          }
        },
        "type": "object",
        "required": ["mode"],
        "title": "ContextConfigRead"
      },
      "ContextMode": {
        "type": "string",
        "enum": ["full"],
        "title": "ContextMode"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JudgeResolutionConfigRead": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "judge_llm",
            "title": "Mode",
            "default": "judge_llm"
          },
          "judge_model": {
            "type": "string",
            "minLength": 1,
            "title": "Judge Model"
          },
          "judge_temperature": {
            "type": "number",
            "title": "Judge Temperature"
          }
        },
        "type": "object",
        "required": ["judge_model", "judge_temperature"],
        "title": "JudgeResolutionConfigRead"
      },
...
      "NoneResolutionConfigRead": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "none",
            "title": "Mode",
            "default": "none"
          }
        },
        "type": "object",
        "title": "NoneResolutionConfigRead"
      },
...
      "ParticipantRead": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/AgentParticipantRead"
          },
          {
            "$ref": "#/components/schemas/UserParticipantRead"
          }
        ],
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "agent": "#/components/schemas/AgentParticipantRead",
            "user": "#/components/schemas/UserParticipantRead"
          }
        }
      },
...
      "ResolutionConfigRead": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/JudgeResolutionConfigRead"
          },
          {
            "$ref": "#/components/schemas/NoneResolutionConfigRead"
          }
        ],
        "discriminator": {
          "propertyName": "mode",
          "mapping": {
            "judge_llm": "#/components/schemas/JudgeResolutionConfigRead",
            "none": "#/components/schemas/NoneResolutionConfigRead"
          }
        }
      },
...
      "SessionConfigRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/ParticipantRead"
            },
            "type": "array",
            "title": "Participants"
          },
          "turn_selection": {
            "$ref": "#/components/schemas/TurnSelectionConfigRead"
          },
          "context": {
            "$ref": "#/components/schemas/ContextConfigRead"
          },
          "validation": {
            "$ref": "#/components/schemas/ValidationConfigRead"
          },
          "resolution": {
            "$ref": "#/components/schemas/ResolutionConfigRead"
          }
        },
        "type": "object",
        "required": [
          "id",
          "prompt",
          "created_at",
          "updated_at",
          "participants",
          "turn_selection",
          "context",
          "validation",
          "resolution"
        ],
        "title": "SessionConfigRead"
      },
...
      "SessionRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "config": {
            "$ref": "#/components/schemas/SessionConfigRead"
          }
        },
        "type": "object",
        "required": ["id", "created_at", "updated_at", "config"],
        "title": "SessionRead"
      },
...
      "TurnSelectionConfigRead": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/TurnSelectionMode"
          }
        },
        "type": "object",
        "required": ["mode"],
        "title": "TurnSelectionConfigRead"
      },
      "TurnSelectionMode": {
        "type": "string",
        "enum": ["round_robin"],
        "title": "TurnSelectionMode"
      },
...
      "UserParticipantRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "const": "user",
            "title": "Type",
            "default": "user"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": ["id", "created_at", "updated_at"],
        "title": "UserParticipantRead"
      },
...
      "ValidationConfigRead": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/ValidationMode"
          }
        },
        "type": "object",
        "required": ["mode"],
        "title": "ValidationConfigRead"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
      "ValidationMode": {
        "type": "string",
        "enum": ["allow_all"],
        "title": "ValidationMode"
      }
    }
  }
}